pyproject.toml
# Recall v7.0 - AI永久记忆系统
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "recall-ai"
version = "7.0.0"
description = "AI永久记忆系统 - 让AI永远不会忘记你说过的每一句话"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Recall Team"}
]
keywords = ["ai", "memory", "recall", "llm", "roleplay", "sillytavern"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # 核心框架
    "pydantic>=2.0",           # 数据模型验证
    
    # NLP处理
    "spacy>=3.5",              # 实体识别
    "jieba>=0.42",             # 中文分词
    
    # LLM调用
    "litellm>=1.0",            # 统一LLM接口
    "openai>=1.0",             # OpenAI SDK
    "httpx>=0.24",             # 异步HTTP
    
    # Web服务
    "fastapi>=0.100",          # HTTP API框架
    "uvicorn>=0.22",           # ASGI服务器
    
    # 系统监控
    "psutil>=5.9",             # 系统资源监控
    "schedule>=1.2",           # 任务调度
    
    # 工具库
    "click>=8.0",              # CLI框架
    "rich>=13.0",              # 终端美化
    "numpy>=1.24",             # 数值计算
    "pybloom-live>=4.0",       # 布隆过滤器
    "pyyaml>=6.0",             # YAML配置（PromptManager）
]

[project.optional-dependencies]
# ====== 新名称（推荐使用）======
# Lite 模式：无向量搜索，最小依赖
lite = []

# Cloud 模式：使用 API Embedding，只需要 FAISS 做向量存储
cloud = [
    "faiss-cpu>=1.7",          # 向量索引
]

# Local 模式：本地 Embedding 模型（默认 GPU 版 PyTorch）
local = [
    "sentence-transformers>=2.2",  # Embedding模型（会自动安装torch GPU版）
    "faiss-cpu>=1.7",              # 向量索引
]

# Local-CPU 模式：本地 Embedding 模型（CPU 版 PyTorch，无需显卡）
# 注意：需要先安装 CPU 版 PyTorch，脚本会自动处理
local-cpu = [
    "sentence-transformers>=2.2",  # Embedding模型
    "faiss-cpu>=1.7",              # 向量索引
]

# Enterprise 模式：Phase 3.5 企业级性能引擎（可选）
enterprise = [
    "faiss-cpu>=1.7",              # 向量索引 (VectorIndexIVF)
    "networkx>=3.0",               # 图分析与社区检测 (CommunityDetector)
    "kuzu>=0.3",                   # 嵌入式图数据库 (KuzuGraphBackend)
]

# Enterprise-CPU 模式：企业级 + CPU 版 PyTorch
enterprise-cpu = [
    "sentence-transformers>=2.2",  # Embedding模型
    "faiss-cpu>=1.7",              # 向量索引 (VectorIndexIVF)
    "networkx>=3.0",               # 图分析与社区检测 (CommunityDetector)
    "kuzu>=0.3",                   # 嵌入式图数据库 (KuzuGraphBackend)
]

# ====== 旧名称（向后兼容）======
# 轻量模式（别名 -> lite）
lightweight = []

# Hybrid 模式（别名 -> cloud）
hybrid = [
    "faiss-cpu>=1.7",          # 向量索引
]

# 完整模式（别名 -> local）
full = [
    "sentence-transformers>=2.2",  # Embedding模型（会自动安装torch）
    "faiss-cpu>=1.7",              # 向量索引
]

# MCP 服务
mcp = ["mcp>=1.0.0", "httpx-sse>=0.4.0", "uvicorn>=0.30.0", "starlette>=0.38.0"]

# LLM 提供商
anthropic = ["anthropic>=0.30.0"]
google = ["google-generativeai>=0.8.0"]
voyage = ["voyageai>=0.3.0"]
all-llm = ["anthropic>=0.30.0", "google-generativeai>=0.8.0", "voyageai>=0.3.0"]

# 重排序后端
reranker = ["cohere>=5.0", "sentence-transformers>=2.2"]

# 开发依赖
dev = ["pytest>=7.0", "black", "ruff"]

[project.scripts]
recall = "recall.cli:main"
recall-mcp = "recall.mcp_server:main"

[project.urls]
Homepage = "https://github.com/recall-ai/recall"
Documentation = "https://github.com/recall-ai/recall"
Repository = "https://github.com/recall-ai/recall"

[tool.setuptools.packages.find]
where = ["."]
include = ["recall*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.css", "*.js"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
# 过滤第三方库的警告
# -W ignore::SyntaxWarning: jieba 在 Python 3.12 下的编译时正则转义警告（第三方库问题，无法通过 filterwarnings 过滤）
addopts = "-W ignore::SyntaxWarning"
filterwarnings = [
    "ignore::UserWarning:jieba.*",
    "ignore::SyntaxWarning:jieba.*",
    "ignore::FutureWarning:transformers.*",
    "ignore::DeprecationWarning",
]
